chore: add react email dev server & fix package npm deps
