+++ .npmrc
@@ -6,3 +6,4 @@ access = public
 enable-pre-post-scripts = true
 legacy-peer-deps=true
 node-linker=hoisted
+save-exact=true
