+++ apps/demo-react-native/package.json
@@ -13,16 +13,16 @@
   "dependencies": {
     "@formbricks/js": "workspace:*",
     "@formbricks/react-native": "workspace:*",
-    "expo": "^51.0.26",
-    "expo-status-bar": "~1.12.1",
-    "react": "^18.2.0",
-    "react-native": "^0.74.4",
+    "expo": "51.0.26",
+    "expo-status-bar": "1.12.1",
+    "react": "18.3.1",
+    "react-native": "0.74.4",
     "react-native-webview": "13.8.6"
   },
   "devDependencies": {
-    "@babel/core": "^7.25.2",
-    "@types/react": "~18.2.79",
-    "typescript": "^5.3.3"
+    "@babel/core": "7.25.2",
+    "@types/react": "18.3.11",
+    "typescript": "5.3.3"
   },
   "private": true
 }
