+++ apps/demo/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@formbricks/js": "workspace:*",
     "@formbricks/ui": "workspace:*",
-    "lucide-react": "^0.418.0",
+    "lucide-react": "0.418.0",
     "next": "14.2.5",
     "react": "18.3.1",
     "react-dom": "18.3.1"
