+++ apps/docs/package.json
@@ -12,62 +12,62 @@
   },
   "browserslist": "defaults, not ie <= 11",
   "dependencies": {
-    "@algolia/autocomplete-core": "^1.17.4",
-    "@calcom/embed-react": "^1.5.1",
+    "@algolia/autocomplete-core": "1.17.4",
+    "@calcom/embed-react": "1.5.1",
     "@docsearch/css": "3",
-    "@docsearch/react": "^3.6.2",
+    "@docsearch/react": "3.6.2",
     "@formbricks/lib": "workspace:*",
     "@formbricks/types": "workspace:*",
     "@formbricks/ui": "workspace:*",
-    "@headlessui/react": "^2.1.9",
-    "@headlessui/tailwindcss": "^0.2.1",
-    "@mapbox/rehype-prism": "^0.9.0",
-    "@mdx-js/loader": "^3.0.1",
-    "@mdx-js/react": "^3.0.1",
+    "@headlessui/react": "2.1.9",
+    "@headlessui/tailwindcss": "0.2.1",
+    "@mapbox/rehype-prism": "0.9.0",
+    "@mdx-js/loader": "3.0.1",
+    "@mdx-js/react": "3.0.1",
     "@next/mdx": "14.2.15",
-    "@paralleldrive/cuid2": "^2.2.2",
-    "@sindresorhus/slugify": "^2.2.1",
-    "@tailwindcss/typography": "^0.5.15",
-    "acorn": "^8.12.1",
-    "autoprefixer": "^10.4.20",
-    "clsx": "^2.1.1",
-    "fast-glob": "^3.3.2",
-    "flexsearch": "^0.7.43",
+    "@paralleldrive/cuid2": "2.2.2",
+    "@sindresorhus/slugify": "2.2.1",
+    "@tailwindcss/typography": "0.5.15",
+    "acorn": "8.12.1",
+    "autoprefixer": "10.4.20",
+    "clsx": "2.1.1",
+    "fast-glob": "3.3.2",
+    "flexsearch": "0.7.43",
     "framer-motion": "11.11.4",
-    "lottie-web": "^5.12.2",
-    "lucide": "^0.451.0",
-    "lucide-react": "^0.451.0",
-    "mdast-util-to-string": "^4.0.0",
-    "mdx-annotations": "^0.1.4",
+    "lottie-web": "5.12.2",
+    "lucide": "0.451.0",
+    "lucide-react": "0.451.0",
+    "mdast-util-to-string": "4.0.0",
+    "mdx-annotations": "0.1.4",
     "next": "14.2.15",
-    "next-plausible": "^3.12.2",
-    "next-seo": "^6.6.0",
-    "next-sitemap": "^4.2.3",
-    "next-themes": "^0.3.0",
-    "node-fetch": "^3.3.2",
-    "prism-react-renderer": "^2.4.0",
-    "prismjs": "^1.29.0",
+    "next-plausible": "3.12.2",
+    "next-seo": "6.6.0",
+    "next-sitemap": "4.2.3",
+    "next-themes": "0.3.0",
+    "node-fetch": "3.3.2",
+    "prism-react-renderer": "2.4.0",
+    "prismjs": "1.29.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-highlight-words": "^0.20.0",
-    "react-markdown": "^9.0.1",
-    "react-responsive-embed": "^2.1.0",
-    "remark": "^15.0.1",
-    "remark-gfm": "^4.0.0",
-    "remark-mdx": "^3.0.1",
-    "schema-dts": "^1.1.2",
-    "sharp": "^0.33.5",
-    "shiki": "^0.14.7",
-    "simple-functional-loader": "^1.2.1",
-    "tailwindcss": "^3.4.13",
-    "unist-util-filter": "^5.0.1",
-    "unist-util-visit": "^5.0.0",
-    "zustand": "^4.5.5"
+    "react-highlight-words": "0.20.0",
+    "react-markdown": "9.0.1",
+    "react-responsive-embed": "2.1.0",
+    "remark": "15.0.1",
+    "remark-gfm": "4.0.0",
+    "remark-mdx": "3.0.1",
+    "schema-dts": "1.1.2",
+    "sharp": "0.33.5",
+    "shiki": "0.14.7",
+    "simple-functional-loader": "1.2.1",
+    "tailwindcss": "3.4.13",
+    "unist-util-filter": "5.0.1",
+    "unist-util-visit": "5.0.0",
+    "zustand": "4.5.5"
   },
   "devDependencies": {
     "@formbricks/config-typescript": "workspace:*",
-    "@types/dompurify": "^3.0.5",
-    "@types/react-highlight-words": "^0.20.0",
+    "@types/dompurify": "3.0.5",
+    "@types/react-highlight-words": "0.20.0",
     "@formbricks/eslint-config": "workspace:*"
   }
 }
