+++ apps/storybook/package.json
@@ -12,30 +12,30 @@
   },
   "dependencies": {
     "@formbricks/ui": "workspace:*",
-    "eslint-plugin-react-refresh": "^0.4.12",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "eslint-plugin-react-refresh": "0.4.12",
+    "react": "18.3.1",
+    "react-dom": "18.3.1"
   },
   "devDependencies": {
-    "@chromatic-com/storybook": "^2.0.2",
+    "@chromatic-com/storybook": "2.0.2",
     "@formbricks/config-typescript": "workspace:*",
-    "@storybook/addon-a11y": "^8.3.5",
-    "@storybook/addon-essentials": "^8.3.5",
-    "@storybook/addon-interactions": "^8.3.5",
-    "@storybook/addon-links": "^8.3.5",
-    "@storybook/addon-onboarding": "^8.3.5",
-    "@storybook/blocks": "^8.3.5",
-    "@storybook/react": "^8.3.5",
-    "@storybook/react-vite": "^8.3.5",
-    "@storybook/test": "^8.3.5",
-    "@typescript-eslint/eslint-plugin": "^8.8.1",
-    "@typescript-eslint/parser": "^8.8.1",
-    "@vitejs/plugin-react": "^4.3.2",
-    "esbuild": "^0.24.0",
-    "eslint-plugin-storybook": "^0.9.0",
-    "prop-types": "^15.8.1",
-    "storybook": "^8.3.5",
-    "tsup": "^8.3.0",
-    "vite": "^5.4.8"
+    "@storybook/addon-a11y": "8.3.5",
+    "@storybook/addon-essentials": "8.3.5",
+    "@storybook/addon-interactions": "8.3.5",
+    "@storybook/addon-links": "8.3.5",
+    "@storybook/addon-onboarding": "8.3.5",
+    "@storybook/blocks": "8.3.5",
+    "@storybook/react": "8.3.5",
+    "@storybook/react-vite": "8.3.5",
+    "@storybook/test": "8.3.5",
+    "@typescript-eslint/eslint-plugin": "8.8.1",
+    "@typescript-eslint/parser": "8.8.1",
+    "@vitejs/plugin-react": "4.3.2",
+    "esbuild": "0.24.0",
+    "eslint-plugin-storybook": "0.9.0",
+    "prop-types": "15.8.1",
+    "storybook": "8.3.5",
+    "tsup": "8.3.0",
+    "vite": "5.4.8"
   }
 }
