+++ apps/web/app/(app)/environments/[environmentId]/surveys/[surveyId]/(analysis)/summary/lib/emailTemplate.tsx
@@ -1,4 +1,4 @@
-import { getPreviewEmailTemplateHtml } from "@formbricks/email/components/survey/preview-email-template";
+import { getPreviewEmailTemplateHtml } from "@formbricks/email/components/preview-email-template";
 import { WEBAPP_URL } from "@formbricks/lib/constants";
 import { getProductByEnvironmentId } from "@formbricks/lib/product/service";
 import { getSurvey } from "@formbricks/lib/survey/service";
@@ -16,7 +16,7 @@ export const getEmailTemplateHtml = async (surveyId: string) => {
 
   const styling = getStyling(product, survey);
   const surveyUrl = WEBAPP_URL + "/s/" + survey.id;
-  const html = getPreviewEmailTemplateHtml(survey, surveyUrl, styling);
+  const html = await getPreviewEmailTemplateHtml(survey, surveyUrl, styling);
   const doctype =
     '<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">';
   const htmlCleaned = html.toString().replace(doctype, "");
