+++ apps/web/package.json
@@ -12,10 +12,10 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@dnd-kit/core": "^6.1.0",
-    "@dnd-kit/modifiers": "^7.0.0",
-    "@dnd-kit/sortable": "^8.0.0",
-    "@dnd-kit/utilities": "^3.2.2",
+    "@dnd-kit/core": "6.1.0",
+    "@dnd-kit/modifiers": "7.0.0",
+    "@dnd-kit/sortable": "8.0.0",
+    "@dnd-kit/utilities": "3.2.2",
     "@formbricks/api": "workspace:*",
     "@formbricks/database": "workspace:*",
     "@formbricks/ee": "workspace:*",
@@ -26,52 +26,52 @@
     "@formbricks/surveys": "workspace:*",
     "@formbricks/types": "workspace:*",
     "@formbricks/ui": "workspace:*",
-    "@hookform/resolvers": "^3.9.0",
-    "@json2csv/node": "^7.0.6",
-    "@paralleldrive/cuid2": "^2.2.2",
-    "@radix-ui/react-collapsible": "^1.1.1",
-    "@react-email/components": "^0.0.25",
-    "@sentry/nextjs": "^8.33.1",
-    "@tanstack/react-table": "^8.20.5",
-    "@vercel/og": "^0.6.3",
-    "@vercel/speed-insights": "^1.0.12",
-    "bcryptjs": "^2.4.3",
-    "dotenv": "^16.4.5",
-    "encoding": "^0.1.13",
-    "file-loader": "^6.2.0",
+    "@hookform/resolvers": "3.9.0",
+    "@json2csv/node": "7.0.6",
+    "@paralleldrive/cuid2": "2.2.2",
+    "@radix-ui/react-collapsible": "1.1.1",
+    "@react-email/components": "0.0.25",
+    "@sentry/nextjs": "8.33.1",
+    "@tanstack/react-table": "8.20.5",
+    "@vercel/og": "0.6.3",
+    "@vercel/speed-insights": "1.0.12",
+    "bcryptjs": "2.4.3",
+    "dotenv": "16.4.5",
+    "encoding": "0.1.13",
+    "file-loader": "6.2.0",
     "framer-motion": "11.11.4",
-    "googleapis": "^144.0.0",
-    "jiti": "^2.3.3",
-    "jsonwebtoken": "^9.0.2",
-    "lodash": "^4.17.21",
-    "lru-cache": "^11.0.1",
-    "lucide-react": "^0.451.0",
-    "mime": "^4.0.4",
+    "googleapis": "144.0.0",
+    "jiti": "2.3.3",
+    "jsonwebtoken": "9.0.2",
+    "lodash": "4.17.21",
+    "lru-cache": "11.0.1",
+    "lucide-react": "0.451.0",
+    "mime": "4.0.4",
     "next": "14.2.15",
-    "next-safe-action": "^7.9.3",
-    "optional": "^0.1.4",
-    "otplib": "^12.0.1",
-    "papaparse": "^5.4.1",
-    "posthog-js": "^1.167.0",
-    "prismjs": "^1.29.0",
+    "next-safe-action": "7.9.3",
+    "optional": "0.1.4",
+    "otplib": "12.0.1",
+    "papaparse": "5.4.1",
+    "posthog-js": "1.167.0",
+    "prismjs": "1.29.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "react-hook-form": "^7.53.0",
-    "react-hot-toast": "^2.4.1",
-    "redis": "^4.7.0",
-    "sharp": "^0.33.5",
-    "ua-parser-js": "^1.0.39",
-    "webpack": "^5.95.0",
-    "xlsx": "^0.18.5"
+    "react-hook-form": "7.53.0",
+    "react-hot-toast": "2.4.1",
+    "redis": "4.7.0",
+    "sharp": "0.33.5",
+    "ua-parser-js": "1.0.39",
+    "webpack": "5.95.0",
+    "xlsx": "0.18.5"
   },
   "devDependencies": {
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
-    "@neshca/cache-handler": "^1.7.3",
-    "@types/bcryptjs": "^2.4.6",
-    "@types/lodash": "^4.17.10",
-    "@types/markdown-it": "^14.1.2",
-    "@types/papaparse": "^5.3.14",
-    "@types/qrcode": "^1.5.5"
+    "@neshca/cache-handler": "1.7.3",
+    "@types/bcryptjs": "2.4.6",
+    "@types/lodash": "4.17.10",
+    "@types/markdown-it": "14.1.2",
+    "@types/papaparse": "5.3.14",
+    "@types/qrcode": "1.5.5"
   }
 }
