+++ package.json
@@ -32,14 +32,14 @@
     "storybook": "turbo run storybook"
   },
   "devDependencies": {
-    "@playwright/test": "^1.45.3",
+    "@playwright/test": "1.45.3",
     "@formbricks/eslint-config": "workspace:*",
-    "eslint": "^8.57.0",
-    "husky": "^9.1.4",
-    "lint-staged": "^15.2.7",
-    "rimraf": "^6.0.1",
-    "tsx": "^4.16.5",
-    "turbo": "^2.0.11"
+    "eslint": "8.57.0",
+    "husky": "9.1.4",
+    "lint-staged": "15.2.7",
+    "rimraf": "6.0.1",
+    "tsx": "4.16.5",
+    "turbo": "2.0.11"
   },
   "lint-staged": {
     "(apps|packages)/**/*.{js,ts,jsx,tsx}": [
@@ -64,7 +64,7 @@
     "showDetails": true
   },
   "dependencies": {
-    "@changesets/cli": "^2.27.7",
-    "playwright": "^1.45.3"
+    "@changesets/cli": "2.27.7",
+    "playwright": "1.45.3"
   }
 }
