+++ packages/api/package.json
@@ -37,11 +37,11 @@
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
     "@formbricks/types": "workspace:*",
-    "@rollup/plugin-inject": "^5.0.5",
-    "buffer": "^6.0.3",
-    "terser": "^5.31.6",
-    "vite": "^5.4.1",
-    "vite-plugin-dts": "^3.9.1",
-    "vite-plugin-node-polyfills": "^0.22.0"
+    "@rollup/plugin-inject": "5.0.5",
+    "buffer": "6.0.3",
+    "terser": "5.31.6",
+    "vite": "5.4.1",
+    "vite-plugin-dts": "3.9.1",
+    "vite-plugin-node-polyfills": "0.22.0"
   }
 }
