+++ packages/config-prettier/package.json
@@ -7,8 +7,8 @@
     "clean": "rimraf node_modules .turbo"
   },
   "devDependencies": {
-    "@trivago/prettier-plugin-sort-imports": "^4.3.0",
-    "prettier": "^3.3.3",
-    "prettier-plugin-tailwindcss": "^0.6.6"
+    "@trivago/prettier-plugin-sort-imports": "4.3.0",
+    "prettier": "3.3.3",
+    "prettier-plugin-tailwindcss": "0.6.6"
   }
 }
