+++ packages/config-typescript/package.json
@@ -8,7 +8,7 @@
   },
   "devDependencies": {
     "@types/node": "22.3.0",
-    "@types/react": "18.3.3",
+    "@types/react": "18.3.11",
     "@types/react-dom": "18.3.0",
     "typescript": "5.4.5"
   }
