+++ packages/database/package.json
@@ -55,20 +55,20 @@
     "data-migration:migrate-survey-types": "ts-node ./data-migrations/20241002123456_migrate_survey_types/data-migration.ts"
   },
   "dependencies": {
-    "@prisma/client": "^5.18.0",
-    "@prisma/extension-accelerate": "^1.1.0",
-    "dotenv-cli": "^7.4.2"
+    "@prisma/client": "5.18.0",
+    "@prisma/extension-accelerate": "1.1.0",
+    "dotenv-cli": "7.4.2"
   },
   "devDependencies": {
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/types": "workspace:*",
-    "@paralleldrive/cuid2": "^2.2.2",
+    "@paralleldrive/cuid2": "2.2.2",
     "@formbricks/eslint-config": "workspace:*",
-    "prisma": "^5.18.0",
-    "prisma-dbml-generator": "^0.12.0",
-    "prisma-json-types-generator": "^3.0.4",
-    "ts-node": "^10.9.2",
-    "zod": "^3.23.8",
-    "zod-prisma": "^0.5.4"
+    "prisma": "5.18.0",
+    "prisma-dbml-generator": "0.12.0",
+    "prisma-json-types-generator": "3.0.4",
+    "ts-node": "10.9.2",
+    "zod": "3.23.8",
+    "zod-prisma": "0.5.4"
   }
 }
