+++ packages/ee/package.json
@@ -11,28 +11,28 @@
     "lint": "eslint --ext .ts,.tsx --fix ."
   },
   "devDependencies": {
-    "@formbricks/config-typescript": "*",
+    "@formbricks/config-typescript": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
-    "@formbricks/lib": "*",
-    "@formbricks/types": "*",
-    "@formbricks/ui": "*",
-    "@types/dompurify": "^3.0.5",
-    "@types/react": "18.3.3"
+    "@formbricks/lib": "workspace:*",
+    "@formbricks/types": "workspace:*",
+    "@formbricks/ui": "workspace:*",
+    "@types/dompurify": "3.0.5",
+    "@types/react": "18.3.11"
   },
   "dependencies": {
     "@formbricks/database": "workspace:*",
     "@formbricks/lib": "workspace:*",
-    "@paralleldrive/cuid2": "^2.2.2",
-    "@radix-ui/react-collapsible": "^1.1.0",
-    "https-proxy-agent": "^7.0.5",
-    "lucide-react": "^0.427.0",
-    "next": "^14.2.5",
-    "next-auth": "^4.24.7",
-    "node-fetch": "^3.3.2",
-    "react-hook-form": "^7.53.0",
-    "react-hot-toast": "^2.4.1",
-    "server-only": "^0.0.1",
-    "stripe": "^16.7.0",
-    "zod": "^3.23.8"
+    "@paralleldrive/cuid2": "2.2.2",
+    "@radix-ui/react-collapsible": "1.1.0",
+    "https-proxy-agent": "7.0.5",
+    "lucide-react": "0.427.0",
+    "next": "14.2.5",
+    "next-auth": "4.24.7",
+    "node-fetch": "3.3.2",
+    "react-hook-form": "7.53.0",
+    "react-hot-toast": "2.4.1",
+    "server-only": "0.0.1",
+    "stripe": "16.7.0",
+    "zod": "3.23.8"
   }
 }
