+++ packages/email/components/email-button.tsx
@@ -13,3 +13,5 @@ export function EmailButton({ label, href }: EmailButtonProps): React.JSX.Elemen
     </Button>
   );
 }
+
+export default EmailButton;
