+++ packages/email/emails/weekly-summary/notification-footer.tsx
