+++ packages/email/package.json
@@ -4,6 +4,7 @@
   "description": "Email package",
   "main": "./index.tsx",
   "scripts": {
+    "dev": "email dev --port 3003",
     "clean": "rimraf .turbo node_modules dist",
     "lint": "eslint --ext .ts,.tsx --fix ."
   },
@@ -12,14 +13,14 @@
     "@formbricks/lib": "workspace:*",
     "@formbricks/types": "workspace:*",
     "@formbricks/ui": "workspace:*",
-    "@react-email/components": "^0.0.22",
-    "@react-email/render": "^0.0.17",
-    "lucide-react": "^0.427.0",
-    "nodemailer": "^6.9.14",
-    "react-email": "^2.1.6"
+    "@react-email/components": "0.0.25",
+    "@react-email/render": "1.0.1",
+    "lucide-react": "0.451.0",
+    "nodemailer": "6.9.15"
   },
   "devDependencies": {
-    "@types/nodemailer": "^6.4.15",
-    "@types/react": "18.3.3"
+    "@types/nodemailer": "6.4.16",
+    "@types/react": "18.3.11",
+    "react-email": "2.1.6"
   }
 }
