+++ packages/email/tsconfig.json
@@ -3,7 +3,7 @@
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "~/*": ["/*"]
+      "@/*": ["./*"]
     },
     "resolveJsonModule": true,
     "strictNullChecks": true
