+++ packages/js-core/package.json
@@ -46,8 +46,8 @@
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/types": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
-    "terser": "^5.31.6",
-    "vite": "^5.4.1",
-    "vite-plugin-dts": "^3.9.1"
+    "terser": "5.31.6",
+    "vite": "5.4.1",
+    "vite-plugin-dts": "3.9.1"
   }
 }
