+++ packages/js/package.json
@@ -44,9 +44,9 @@
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
     "@formbricks/types": "workspace:*",
-    "terser": "^5.31.6",
-    "vite": "^5.4.1",
-    "vite-plugin-dts": "^3.9.1"
+    "terser": "5.31.6",
+    "vite": "5.4.1",
+    "vite-plugin-dts": "3.9.1"
   },
   "peerDependencies": {
     "zod": "3.x"
