+++ packages/lib/package.json
@@ -20,31 +20,31 @@
     "@formbricks/api": "*",
     "@formbricks/database": "*",
     "@formbricks/types": "*",
-    "@paralleldrive/cuid2": "^2.2.2",
-    "@t3-oss/env-nextjs": "^0.11.0",
-    "@ungap/structured-clone": "^1.2.0",
-    "aws-crt": "^1.21.3",
-    "date-fns": "^3.6.0",
-    "jsonwebtoken": "^9.0.2",
-    "markdown-it": "^14.1.0",
-    "mime-types": "^2.1.35",
-    "nanoid": "^5.0.7",
-    "next-auth": "^4.24.7",
-    "posthog-node": "^4.1.0",
-    "qrcode": "^1.5.4",
-    "server-only": "^0.0.1",
-    "superjson": "^2.2.1",
-    "tailwind-merge": "^2.5.2"
+    "@paralleldrive/cuid2": "2.2.2",
+    "@t3-oss/env-nextjs": "0.11.0",
+    "@ungap/structured-clone": "1.2.0",
+    "aws-crt": "1.21.3",
+    "date-fns": "3.6.0",
+    "jsonwebtoken": "9.0.2",
+    "markdown-it": "14.1.0",
+    "mime-types": "2.1.35",
+    "nanoid": "5.0.7",
+    "next-auth": "4.24.7",
+    "posthog-node": "4.1.0",
+    "qrcode": "1.5.4",
+    "server-only": "0.0.1",
+    "superjson": "2.2.1",
+    "tailwind-merge": "2.5.2"
   },
   "devDependencies": {
     "@formbricks/config-typescript": "workspace:*",
-    "@types/jsonwebtoken": "^9.0.6",
-    "@types/mime-types": "^2.1.4",
-    "@types/ungap__structured-clone": "^1.2.0",
-    "dotenv": "^16.4.5",
+    "@types/jsonwebtoken": "9.0.6",
+    "@types/mime-types": "2.1.4",
+    "@types/ungap__structured-clone": "1.2.0",
+    "dotenv": "16.4.5",
     "@formbricks/eslint-config": "workspace:*",
-    "ts-node": "^10.9.2",
-    "vitest": "^2.0.5",
-    "vitest-mock-extended": "^2.0.0"
+    "ts-node": "10.9.2",
+    "vitest": "2.0.5",
+    "vitest-mock-extended": "2.0.0"
   }
 }
