--- packages/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "react-select": "^5.8.0"
-  }
-}
