+++ packages/surveys/package.json
@@ -41,21 +41,21 @@
     "@formbricks/eslint-config": "workspace:*",
     "@formbricks/lib": "workspace:*",
     "@formbricks/types": "workspace:*",
-    "@preact/preset-vite": "^2.9.0",
-    "autoprefixer": "^10.4.20",
+    "@preact/preset-vite": "2.9.0",
+    "autoprefixer": "10.4.20",
     "concurrently": "8.2.2",
-    "isomorphic-dompurify": "^2.14.0",
-    "postcss": "^8.4.41",
-    "preact": "^10.23.2",
-    "react-date-picker": "^11.0.0",
+    "isomorphic-dompurify": "2.14.0",
+    "postcss": "8.4.41",
+    "preact": "10.23.2",
+    "react-date-picker": "11.0.0",
     "serve": "14.2.3",
-    "tailwindcss": "^3.4.10",
-    "terser": "^5.31.6",
-    "vite": "^5.4.1",
-    "vite-plugin-dts": "^3.9.1",
-    "vite-tsconfig-paths": "^5.0.1"
+    "tailwindcss": "3.4.10",
+    "terser": "5.31.6",
+    "vite": "5.4.1",
+    "vite-plugin-dts": "3.9.1",
+    "vite-tsconfig-paths": "5.0.1"
   },
   "dependencies": {
-    "@formkit/auto-animate": "^0.8.2"
+    "@formkit/auto-animate": "0.8.2"
   }
 }
