+++ packages/ui/package.json
@@ -12,50 +12,50 @@
     "@formbricks/config-typescript": "workspace:*",
     "@formbricks/eslint-config": "workspace:*",
     "@formbricks/types": "workspace:*",
-    "concurrently": "^8.2.2",
-    "postcss": "^8.4.41",
+    "concurrently": "8.2.2",
+    "postcss": "8.4.41",
     "react": "18.3.1"
   },
   "dependencies": {
     "@formbricks/database": "workspace:*",
     "@formbricks/lib": "workspace:*",
     "@formbricks/surveys": "workspace:*",
-    "@lexical/code": "^0.17.0",
-    "@lexical/link": "^0.17.0",
-    "@lexical/list": "^0.17.0",
-    "@lexical/markdown": "^0.17.0",
-    "@lexical/react": "^0.17.0",
-    "@lexical/rich-text": "^0.17.0",
-    "@lexical/table": "^0.17.0",
-    "@radix-ui/react-accordion": "^1.2.0",
-    "@radix-ui/react-checkbox": "^1.1.1",
-    "@radix-ui/react-dialog": "^1.1.1",
-    "@radix-ui/react-dropdown-menu": "^2.1.1",
-    "@radix-ui/react-label": "^2.1.0",
-    "@radix-ui/react-popover": "^1.1.1",
-    "@radix-ui/react-radio-group": "^1.2.0",
-    "@radix-ui/react-select": "^2.1.1",
-    "@radix-ui/react-slider": "^1.2.0",
-    "@radix-ui/react-slot": "^1.1.0",
-    "@radix-ui/react-switch": "^1.1.0",
-    "@radix-ui/react-tooltip": "^1.1.2",
-    "@tailwindcss/forms": "^0.5.9",
-    "@tailwindcss/typography": "^0.5.13",
-    "autoprefixer": "^10.4.20",
-    "boring-avatars": "^1.10.2",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.1",
-    "cmdk": "^1.0.0",
-    "lexical": "^0.17.0",
-    "lucide-react": "^0.427.0",
-    "mime": "^4.0.4",
-    "react-colorful": "^5.6.1",
-    "react-confetti": "^6.1.0",
-    "react-day-picker": "^9.0.8",
-    "react-hot-toast": "^2.4.1",
-    "react-radio-group": "^3.0.3",
-    "react-use": "^17.5.1",
-    "tailwind-merge": "^2.5.2",
-    "tailwindcss": "^3.4.13"
+    "@lexical/code": "0.17.0",
+    "@lexical/link": "0.17.0",
+    "@lexical/list": "0.17.0",
+    "@lexical/markdown": "0.17.0",
+    "@lexical/react": "0.17.0",
+    "@lexical/rich-text": "0.17.0",
+    "@lexical/table": "0.17.0",
+    "@radix-ui/react-accordion": "1.2.0",
+    "@radix-ui/react-checkbox": "1.1.1",
+    "@radix-ui/react-dialog": "1.1.1",
+    "@radix-ui/react-dropdown-menu": "2.1.1",
+    "@radix-ui/react-label": "2.1.0",
+    "@radix-ui/react-popover": "1.1.1",
+    "@radix-ui/react-radio-group": "1.2.0",
+    "@radix-ui/react-select": "2.1.1",
+    "@radix-ui/react-slider": "1.2.0",
+    "@radix-ui/react-slot": "1.1.0",
+    "@radix-ui/react-switch": "1.1.0",
+    "@radix-ui/react-tooltip": "1.1.2",
+    "@tailwindcss/forms": "0.5.9",
+    "@tailwindcss/typography": "0.5.13",
+    "autoprefixer": "10.4.20",
+    "boring-avatars": "1.10.2",
+    "class-variance-authority": "0.7.0",
+    "clsx": "2.1.1",
+    "cmdk": "1.0.0",
+    "lexical": "0.17.0",
+    "lucide-react": "0.427.0",
+    "mime": "4.0.4",
+    "react-colorful": "5.6.1",
+    "react-confetti": "6.1.0",
+    "react-day-picker": "9.0.8",
+    "react-hot-toast": "2.4.1",
+    "react-radio-group": "3.0.3",
+    "react-use": "17.5.1",
+    "tailwind-merge": "2.5.2",
+    "tailwindcss": "3.4.13"
   }
 }
